refactor: rename components for consistency and improve structure
