+++ components/sections/blog/blog-client.tsx
@@ -4,9 +4,9 @@ import { useRouter, useSearchParams } from "next/navigation";
 import { useEffect, useState } from "react";
 import { LatestBlogHero } from "@/components/sections/blog/latest-blog-hero";
 import { Post } from "@/sanity/schemaTypes/postType";
-import { Tag } from "@/components/blog/tag";
+import { Tag } from "@/components/blog/Tag";
 import { Input } from "@/components/ui/input";
-import { ArticleCard } from "@/components/blog/article-card";
+import { ArticleCard } from "@/components/blog/ArticleCard";
 import { useInView } from "react-intersection-observer";
 import { motion } from "motion/react";
 import { getPosts } from "@/lib/blog";
